feat: Implement macOS Screen Capture Kit support for Phase 1
- Add modern macOS screen capture implementation using Core Graphics
- Support for screen capture and window capture functionality
- Implement ScreenCaptureBackend and WindowCaptureBackend traits
- Add comprehensive helper functions for display/window enumeration
- Phase 1: Basic implementation using CGDisplayCreateImage/CGWindowListCreateImage
- Thread-safe implementation without Screen Capture Kit streaming (Phase 2)
- All 91 tests passing, full clippy compliance
- Refactor macOS capture module to use proper directory structure

Technical details:
- Uses Core Graphics APIs for compatibility with older macOS versions
- Placeholder image data conversion with test patterns for validation
- Proper resource cleanup and error handling
- Cross-platform abstraction layer maintained
