+++ crates/constellation-nodes/Cargo.toml
@@ -45,6 +45,8 @@ core-foundation = "0.9"
 core-media = "0.1"
 cocoa = "0.25"
 objc = "0.2"
+objc-foundation = "0.1"
+block = "0.1"
 
 [target.'cfg(target_os = "linux")'.dependencies]
 x11 = "2.21"
