--- crates/constellation-nodes/src/capture/macos.rs
@@ -1,241 +1,0 @@
-/*
- * Constellation Studio - Professional Real-time Video Processing
- * Copyright (c) 2025 MACHIKO LAB
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see <https://www.gnu.org/licenses/>.
- */
-
-use super::{ScreenCaptureBackend, WindowCaptureBackend, WindowInfo};
-#[cfg(target_os = "macos")]
-use anyhow::Result;
-use constellation_core::{VideoFormat, VideoFrame};
-
-use core_graphics::{
-    display::{CGDisplayBounds, CGMainDisplayID},
-    window::CGWindowID,
-};
-
-use std::ptr;
-
-pub struct MacOSScreenCapture {
-    display_id: u32,
-    capture_cursor: bool,
-    width: u32,
-    height: u32,
-}
-
-impl ScreenCaptureBackend for MacOSScreenCapture {
-    fn new(display_id: u32, capture_cursor: bool) -> Result<Self> {
-        let cg_display = unsafe {
-            if display_id == 0 {
-                CGMainDisplayID()
-            } else {
-                // For secondary displays, we'll need to enumerate them
-                CGMainDisplayID() // Simplified for now
-            }
-        };
-
-        let bounds = unsafe { CGDisplayBounds(cg_display) };
-        let width = bounds.size.width as u32;
-        let height = bounds.size.height as u32;
-
-        Ok(Self {
-            display_id,
-            capture_cursor,
-            width,
-            height,
-        })
-    }
-
-    fn capture_frame(&mut self) -> Result<VideoFrame> {
-        let cg_display = unsafe {
-            // TODO: Support multiple displays
-            CGMainDisplayID()
-        };
-
-        // Create a screenshot using CGDisplayCreateImage
-        let image = unsafe { core_graphics::display::CGDisplayCreateImage(cg_display) };
-
-        if image.is_null() {
-            return Err(anyhow::anyhow!("Failed to capture screen image"));
-        }
-
-        let frame_data = self.convert_cg_image_to_frame_data(image)?;
-
-        // The image will be automatically released when it goes out of scope
-
-        Ok(VideoFrame {
-            width: self.width,
-            height: self.height,
-            format: VideoFormat::Rgba8,
-            data: frame_data,
-        })
-    }
-
-    fn get_display_count() -> Result<u32> {
-        get_display_count()
-    }
-
-    fn get_display_bounds(&self, display_id: u32) -> Result<(u32, u32, u32, u32)> {
-        let cg_display = unsafe {
-            // TODO: Support multiple displays
-            CGMainDisplayID()
-        };
-
-        let bounds = unsafe { CGDisplayBounds(cg_display) };
-        Ok((
-            bounds.origin.x as u32,
-            bounds.origin.y as u32,
-            bounds.size.width as u32,
-            bounds.size.height as u32,
-        ))
-    }
-}
-
-impl MacOSScreenCapture {
-    fn convert_cg_image_to_frame_data(
-        &self,
-        _image: *mut core_graphics::sys::CGImage,
-    ) -> Result<Vec<u8>> {
-        // Simplified placeholder implementation
-        // In a real implementation, we would properly extract the image data
-        let size = (self.width * self.height * 4) as usize;
-        Ok(vec![0u8; size])
-    }
-}
-
-pub struct MacOSWindowCapture {
-    window_id: CGWindowID,
-    width: u32,
-    height: u32,
-}
-
-impl WindowCaptureBackend for MacOSWindowCapture {
-    fn new(window_id: u64) -> Result<Self> {
-        let cg_window_id = window_id as CGWindowID;
-        let (width, height) = get_window_dimensions(cg_window_id)?;
-
-        Ok(Self {
-            window_id: cg_window_id,
-            width,
-            height,
-        })
-    }
-
-    fn new_by_title(title: &str) -> Result<Self> {
-        let window_id = find_window_by_title(title)?;
-        let (width, height) = get_window_dimensions(window_id)?;
-
-        Ok(Self {
-            window_id,
-            width,
-            height,
-        })
-    }
-
-    fn capture_frame(&mut self) -> Result<VideoFrame> {
-        // Create window capture using CGWindowListCreateImage
-        let image = unsafe {
-            core_graphics::window::CGWindowListCreateImage(
-                core_graphics::geometry::CGRect::new(
-                    &core_graphics::geometry::CGPoint::new(0.0, 0.0),
-                    &core_graphics::geometry::CGSize::new(self.width as f64, self.height as f64),
-                ),
-                core_graphics::window::kCGWindowListOptionIncludingWindow,
-                self.window_id,
-                core_graphics::window::kCGWindowImageDefault,
-            )
-        };
-
-        if image.is_null() {
-            return Err(anyhow::anyhow!("Failed to capture window image"));
-        }
-
-        let frame_data = self.convert_cg_image_to_frame_data(image)?;
-
-        // Image will be automatically released
-
-        Ok(VideoFrame {
-            width: self.width,
-            height: self.height,
-            format: VideoFormat::Rgba8,
-            data: frame_data,
-        })
-    }
-
-    fn get_window_list() -> Result<Vec<WindowInfo>> {
-        get_window_list()
-    }
-
-    fn get_window_bounds(&self) -> Result<(u32, u32, u32, u32)> {
-        get_window_bounds(self.window_id)
-    }
-}
-
-impl MacOSWindowCapture {
-    fn convert_cg_image_to_frame_data(
-        &self,
-        _image: *mut core_graphics::sys::CGImage,
-    ) -> Result<Vec<u8>> {
-        // Simplified placeholder implementation
-        let size = (self.width * self.height * 4) as usize;
-        Ok(vec![0u8; size])
-    }
-}
-
-// Helper functions
-pub fn get_display_count() -> Result<u32> {
-    unsafe {
-        let mut display_count = 0u32;
-        let result =
-            core_graphics::display::CGGetActiveDisplayList(0, ptr::null_mut(), &mut display_count);
-
-        if result == core_graphics::base::kCGErrorSuccess {
-            Ok(display_count)
-        } else {
-            Err(anyhow::anyhow!("Failed to get display count"))
-        }
-    }
-}
-
-fn get_window_dimensions(_window_id: CGWindowID) -> Result<(u32, u32)> {
-    // Simplified placeholder implementation
-    Ok((800, 600))
-}
-
-fn get_window_bounds(_window_id: CGWindowID) -> Result<(u32, u32, u32, u32)> {
-    // Simplified placeholder implementation
-    Ok((0, 0, 800, 600))
-}
-
-fn get_window_info(_window_id: CGWindowID) -> Result<()> {
-    // Simplified placeholder - would need proper Core Foundation handling
-    Err(anyhow::anyhow!("get_window_info not implemented"))
-}
-
-fn find_window_by_title(_title: &str) -> Result<CGWindowID> {
-    // Simplified placeholder
-    Ok(1)
-}
-
-pub fn get_window_list() -> Result<Vec<WindowInfo>> {
-    // Simplified placeholder implementation
-    // In a real implementation, this would enumerate actual windows
-    Ok(vec![WindowInfo {
-        id: 1,
-        title: "Test Window".to_string(),
-        process_name: "Test Process".to_string(),
-        bounds: (0, 0, 800, 600),
-    }])
-}
