+++ crates/constellation-nodes/src/capture/macos/macos_screen_capture_kit.rs
@@ -1,0 +1,361 @@
+/*
+ * Constellation Studio - Professional Real-time Video Processing
+ * Copyright (c) 2025 MACHIKO LAB
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+use crate::capture::{ScreenCaptureBackend, WindowCaptureBackend, WindowInfo};
+use anyhow::Result;
+use constellation_core::{VideoFormat, VideoFrame};
+use core_graphics::display::{CGDisplayBounds, CGMainDisplayID};
+
+/// Modern macOS Screen Capture Kit implementation
+/// Uses Screen Capture Kit API (macOS 12.3+) for optimal performance
+/// Phase 1: Basic implementation using CGDisplayCreateImage for compatibility
+pub struct ScreenCaptureKitCapture {
+    display_id: u32,
+    capture_cursor: bool,
+    width: u32,
+    height: u32,
+    // Phase 1: Remove Screen Capture Kit stream for now - will be implemented in Phase 2
+    // capture_stream: Option<Arc<Mutex<*mut Object>>>,
+}
+
+impl ScreenCaptureBackend for ScreenCaptureKitCapture {
+    fn new(display_id: u32, capture_cursor: bool) -> Result<Self> {
+        let cg_display = unsafe {
+            // For Phase 1, we only support the main display
+            // Multi-display support will be added in Phase 2
+            CGMainDisplayID()
+        };
+
+        let bounds = unsafe { CGDisplayBounds(cg_display) };
+        let width = bounds.size.width as u32;
+        let height = bounds.size.height as u32;
+
+        let capture = Self {
+            display_id,
+            capture_cursor,
+            width,
+            height,
+            // Phase 1: Remove Screen Capture Kit stream for now
+            // capture_stream: None,
+        };
+
+        // Phase 1: Skip Screen Capture Kit initialization for now
+        // capture.initialize_screen_capture_kit()?;
+
+        Ok(capture)
+    }
+
+    fn capture_frame(&mut self) -> Result<VideoFrame> {
+        // For now, fallback to CGDisplayCreateImage for compatibility
+        // Full Screen Capture Kit implementation would use streaming
+        self.capture_frame_fallback()
+    }
+
+    fn get_display_count() -> Result<u32> {
+        get_display_count()
+    }
+
+    fn get_display_bounds(&self, _display_id: u32) -> Result<(u32, u32, u32, u32)> {
+        let cg_display = unsafe {
+            // For Phase 1, we only support the main display
+            // Multi-display support will be added in Phase 2
+            CGMainDisplayID()
+        };
+
+        let bounds = unsafe { CGDisplayBounds(cg_display) };
+        Ok((
+            bounds.origin.x as u32,
+            bounds.origin.y as u32,
+            bounds.size.width as u32,
+            bounds.size.height as u32,
+        ))
+    }
+}
+
+impl ScreenCaptureKitCapture {
+    // Phase 1: Remove Screen Capture Kit initialization methods
+    // These will be implemented in Phase 2 with proper thread safety
+
+    fn capture_frame_fallback(&mut self) -> Result<VideoFrame> {
+        let cg_display = unsafe {
+            // For Phase 1, we only support the main display
+            // Multi-display support will be added in Phase 2
+            CGMainDisplayID()
+        };
+
+        // Create a screenshot using CGDisplayCreateImage
+        let image = unsafe { core_graphics::display::CGDisplayCreateImage(cg_display) };
+
+        if image.is_null() {
+            return Err(anyhow::anyhow!("Failed to capture screen image"));
+        }
+
+        let frame_data = self.convert_cg_image_to_frame_data(image)?;
+
+        // The image will be automatically released when it goes out of scope
+        // Note: CGImageRelease is handled by the CGImage wrapper
+
+        Ok(VideoFrame {
+            width: self.width,
+            height: self.height,
+            format: VideoFormat::Rgba8,
+            data: frame_data,
+        })
+    }
+
+    fn convert_cg_image_to_frame_data(
+        &self,
+        _image: *mut core_graphics::sys::CGImage,
+    ) -> Result<Vec<u8>> {
+        // Simplified placeholder implementation
+        // In a real implementation, we would properly extract the image data
+        // using CGImageGetDataProvider and CGDataProviderCopyData
+        let size = (self.width * self.height * 4) as usize;
+
+        // Create a test pattern for now
+        let mut buffer = vec![0u8; size];
+        for i in (0..size).step_by(4) {
+            buffer[i] = 255; // Red
+            buffer[i + 1] = 0; // Green
+            buffer[i + 2] = 0; // Blue
+            buffer[i + 3] = 255; // Alpha
+        }
+
+        Ok(buffer)
+    }
+}
+
+impl Drop for ScreenCaptureKitCapture {
+    fn drop(&mut self) {
+        // Phase 1: No resources to clean up
+        // Clean up Screen Capture Kit resources will be implemented in Phase 2
+        tracing::debug!(
+            "Screen capture resources cleaned up for display {}",
+            self.display_id
+        );
+    }
+}
+
+// Helper functions
+pub fn get_display_count() -> Result<u32> {
+    use std::ptr;
+
+    unsafe {
+        let mut display_count = 0u32;
+        let result =
+            core_graphics::display::CGGetActiveDisplayList(0, ptr::null_mut(), &mut display_count);
+
+        if result == core_graphics::base::kCGErrorSuccess {
+            Ok(display_count)
+        } else {
+            Err(anyhow::anyhow!("Failed to get display count"))
+        }
+    }
+}
+
+pub fn get_display_list() -> Result<Vec<u32>> {
+    use std::ptr;
+
+    let display_count = get_display_count()?;
+    let mut displays = vec![0u32; display_count as usize];
+
+    unsafe {
+        let result = core_graphics::display::CGGetActiveDisplayList(
+            display_count,
+            displays.as_mut_ptr(),
+            ptr::null_mut(),
+        );
+
+        if result == core_graphics::base::kCGErrorSuccess {
+            Ok(displays)
+        } else {
+            Err(anyhow::anyhow!("Failed to get display list"))
+        }
+    }
+}
+
+/// Window capture implementation using Screen Capture Kit
+/// Phase 1: Basic implementation using CGWindowListCreateImage
+pub struct ScreenCaptureKitWindowCapture {
+    window_id: u32,
+    width: u32,
+    height: u32,
+    // Phase 1: Remove Screen Capture Kit stream for now
+    // capture_stream: Option<Arc<Mutex<*mut Object>>>,
+}
+
+impl WindowCaptureBackend for ScreenCaptureKitWindowCapture {
+    fn new(window_id: u64) -> Result<Self> {
+        let window_id = window_id as u32;
+        let (width, height) = get_window_dimensions(window_id)?;
+
+        let capture = Self {
+            window_id,
+            width,
+            height,
+            // Phase 1: Remove Screen Capture Kit stream for now
+            // capture_stream: None,
+        };
+
+        // Phase 1: Skip Screen Capture Kit initialization for now
+        // capture.initialize_window_capture_kit()?;
+
+        Ok(capture)
+    }
+
+    fn new_by_title(title: &str) -> Result<Self> {
+        let window_id = find_window_by_title(title)?;
+        let (width, height) = get_window_dimensions(window_id)?;
+
+        let capture = Self {
+            window_id,
+            width,
+            height,
+            // Phase 1: Remove Screen Capture Kit stream for now
+            // capture_stream: None,
+        };
+
+        // Phase 1: Skip Screen Capture Kit initialization for now
+        // capture.initialize_window_capture_kit()?;
+
+        Ok(capture)
+    }
+
+    fn capture_frame(&mut self) -> Result<VideoFrame> {
+        // For now, fallback to CGWindowListCreateImage
+        self.capture_frame_fallback()
+    }
+
+    fn get_window_list() -> Result<Vec<WindowInfo>> {
+        get_window_list()
+    }
+
+    fn get_window_bounds(&self) -> Result<(u32, u32, u32, u32)> {
+        get_window_bounds(self.window_id)
+    }
+}
+
+impl ScreenCaptureKitWindowCapture {
+    // Phase 1: Remove Screen Capture Kit initialization methods
+    // These will be implemented in Phase 2 with proper thread safety
+
+    fn capture_frame_fallback(&mut self) -> Result<VideoFrame> {
+        // Create window capture using CGWindowListCreateImage
+        let image = unsafe {
+            core_graphics::window::CGWindowListCreateImage(
+                core_graphics::geometry::CGRect::new(
+                    &core_graphics::geometry::CGPoint::new(0.0, 0.0),
+                    &core_graphics::geometry::CGSize::new(self.width as f64, self.height as f64),
+                ),
+                core_graphics::window::kCGWindowListOptionIncludingWindow,
+                self.window_id,
+                core_graphics::window::kCGWindowImageDefault,
+            )
+        };
+
+        if image.is_null() {
+            return Err(anyhow::anyhow!("Failed to capture window image"));
+        }
+
+        let frame_data = self.convert_cg_image_to_frame_data(image)?;
+
+        // Release the image
+        // Note: CGImageRelease is handled by the CGImage wrapper
+
+        Ok(VideoFrame {
+            width: self.width,
+            height: self.height,
+            format: VideoFormat::Rgba8,
+            data: frame_data,
+        })
+    }
+
+    fn convert_cg_image_to_frame_data(
+        &self,
+        _image: *mut core_graphics::sys::CGImage,
+    ) -> Result<Vec<u8>> {
+        // Simplified placeholder implementation
+        // In a real implementation, we would properly extract the image data
+        let size = (self.width * self.height * 4) as usize;
+
+        // Create a test pattern for now (blue for window capture)
+        let mut buffer = vec![0u8; size];
+        for i in (0..size).step_by(4) {
+            buffer[i] = 0; // Red
+            buffer[i + 1] = 0; // Green
+            buffer[i + 2] = 255; // Blue
+            buffer[i + 3] = 255; // Alpha
+        }
+
+        Ok(buffer)
+    }
+}
+
+impl Drop for ScreenCaptureKitWindowCapture {
+    fn drop(&mut self) {
+        // Phase 1: No resources to clean up
+        // Clean up Screen Capture Kit resources will be implemented in Phase 2
+        tracing::debug!(
+            "Window capture resources cleaned up for window {}",
+            self.window_id
+        );
+    }
+}
+
+// Helper functions
+fn get_window_dimensions(window_id: u32) -> Result<(u32, u32)> {
+    // Get window bounds using Core Graphics
+    let bounds = get_window_bounds(window_id)?;
+    Ok((bounds.2, bounds.3)) // width, height
+}
+
+fn get_window_bounds(window_id: u32) -> Result<(u32, u32, u32, u32)> {
+    // Simplified placeholder implementation
+    // In a real implementation, we would use CGWindowListCopyWindowInfo
+    // to get actual window bounds
+    tracing::debug!("Getting bounds for window {}", window_id);
+    Ok((0, 0, 800, 600))
+}
+
+fn find_window_by_title(title: &str) -> Result<u32> {
+    // Simplified placeholder implementation
+    // In a real implementation, we would search through the window list
+    tracing::debug!("Looking for window with title: {}", title);
+
+    // Return a dummy window ID for now
+    Ok(1)
+}
+
+pub fn get_window_list() -> Result<Vec<WindowInfo>> {
+    // Simplified placeholder implementation
+    // In a real implementation, we would enumerate actual windows using CGWindowListCopyWindowInfo
+    Ok(vec![
+        WindowInfo {
+            id: 1,
+            title: "Test Window 1".to_string(),
+            process_name: "Test Process".to_string(),
+            bounds: (0, 0, 800, 600),
+        },
+        WindowInfo {
+            id: 2,
+            title: "Test Window 2".to_string(),
+            process_name: "Another Process".to_string(),
+            bounds: (100, 100, 1024, 768),
+        },
+    ])
+}
